refactored file names to bring in line with google best pratices. Working on controlling exposure of only methods/classes that we want to be part of the public interface.
